Bounce System
Enemy uses 8-dimentional direction to implmenent bounce.

It checks for node that 45 and -45 degree nearly of current, that he want to walk in.

If both of them it borders (opposite state or non nodes at all) - bounce in opposite direction.

If only one of 45 / -45 degree neighbour is avaliable - rotates his move direction to 90 / -90 degree accordingly.